# internal/docs/swagger.yaml
basePath: /
definitions:
...
  api.createUserRequest:
    properties:
      email:
        type: string
      isAdmin:
        type: boolean
    type: object
  api.disableRequest:
    properties:
      code:
        type: string
      password:
        type: string
    type: object
...
  api.enrolFinishRequest:
    properties:
      code:
        type: string
      enrolmentId:
        type: string
    type: object
  api.errorBody:
    properties:
      error:
        type: string
    type: object
...
  api.passwordRequest:
    properties:
      password:
        type: string
    type: object
...
  domain.InstanceSettings:
    properties:
      allowRegister:
        type: boolean
      updatedAt:
        type: string
      usageReportAtMinute:
        description: UsageReportAtMinute is the local wall-clock time of day the report
          runs.
        type: integer
      usageReportRunAt:
        type: string
      usageReportTimeZone:
        type: string
    type: object
...
  service.Report:
    properties:
      accounts:
        items:
          $ref: '#/definitions/service.AccountUsage'
        type: array
      generatedAt:
        type: string
      limits:
        $ref: '#/definitions/service.Quotas'
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      usageReportAtMinute:
        type: integer
    type: object
...
info:
  contact: {}
  description: 'A GTD web application: actions, contexts, projects and the tickler.'
  title: gotracks API
  version: "1.0"
paths:
  /api/v1/2fa:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Two-factor status
      tags:
      - twofactor
  /api/v1/2fa/disable:
    post:
      parameters:
      - description: Current password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.disableRequest'
      responses:
        "204":
          description: disabled
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Disable two-factor
      tags:
      - twofactor
  /api/v1/2fa/enrol/begin:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Begin two-factor enrolment
      tags:
      - twofactor
  /api/v1/2fa/enrol/finish:
    post:
      parameters:
      - description: Confirmation code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.enrolFinishRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Finish two-factor enrolment
      tags:
      - twofactor
  /api/v1/2fa/recovery/regenerate:
    post:
      parameters:
      - description: Current password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.passwordRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Regenerate recovery codes
      tags:
      - twofactor
  /api/v1/admin/reports/usage:
    get:
      parameters:
      - description: Search
        in: query
        name: q
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Report'
      security:
      - BearerAuth: []
      summary: Get the instance usage report
      tags:
      - admin
  /api/v1/admin/reports/usage/run:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Rebuild the usage report
      tags:
      - admin
  /api/v1/admin/settings:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.InstanceSettings'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Get instance settings
      tags:
      - admin
    put:
      parameters:
      - description: Settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.instanceSettingsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.InstanceSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Update instance settings
      tags:
      - admin
  /api/v1/admin/users:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.adminUser'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - admin
    post:
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - admin
  /api/v1/admin/users/{id}:
    delete:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: deleted
        "409":
          description: last admin or self-delete
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - admin
    put:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to change
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - admin
  /api/v1/admin/users/{id}/2fa/reset:
    post:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: reset
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Reset a user's two-factor
      tags:
      - admin
  /api/v1/admin/users/{id}/invitation:
    post:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: sent
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Resend a user invitation
      tags:
      - admin
  /api/v1/admin/users/{id}/usage:
    get:
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.Usage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Get a user's quota usage
      tags:
      - admin
  /api/v1/attachments:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AttachmentWithTodo'
            type: array
      security:
      - BearerAuth: []
      summary: List all attachments
      tags:
      - attachments
  /api/v1/attachments/{id}:
    delete:
      parameters:
      - description: Attachment id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: deleted
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Delete an attachment
      tags:
      - attachments
    get:
      parameters:
      - description: Attachment id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Download an attachment
      tags:
      - attachments
  /api/v1/auth/2fa/verify:
    post:
      parameters:
      - description: Challenge id and code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.twoFactorVerifyRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Complete a two-factor sign-in
      tags:
      - auth
  /api/v1/auth/account/deletion/confirm:
    post:
      parameters:
      - description: Mailed deletion token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.tokenRequest'
      responses:
        "204":
          description: account deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "409":
          description: last administrator
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Confirm permanent account deletion
      tags:
      - auth
  /api/v1/auth/email/change/confirm:
    post:
      parameters:
      - description: Mailed email-change token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.tokenRequest'
      responses:
        "204":
          description: email changed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Confirm a new email address
      tags:
      - auth
  /api/v1/auth/email/resend:
    post:
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.emailRequest'
      responses:
        "204":
          description: accepted
      summary: Resend the verification email
      tags:
      - auth
  /api/v1/auth/email/verify:
    post:
      parameters:
      - description: Mailed token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.tokenRequest'
      responses:
        "204":
          description: verified
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Verify an email address
      tags:
      - auth
  /api/v1/auth/invitation/accept:
    post:
      parameters:
      - description: Invitation token and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.resetPasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Accept a user invitation
      tags:
      - auth
  /api/v1/auth/login:
    post:
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Sign in with email and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      responses:
        "204":
          description: signed out
      summary: Sign out (revoke a refresh token)
      tags:
      - auth
  /api/v1/auth/passkey/login/begin:
    post:
      parameters:
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.passkeyLoginBeginRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Begin passkey sign-in
      tags:
      - passkeys
  /api/v1/auth/passkey/login/finish:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Finish passkey sign-in
      tags:
      - passkeys
...
  /api/v1/auth/password/forgot:
    post:
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.emailRequest'
      responses:
        "204":
          description: accepted
      summary: Request a password-reset email
      tags:
      - auth
  /api/v1/auth/password/reset:
    post:
      parameters:
      - description: Token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.resetPasswordRequest'
      responses:
        "204":
          description: password changed
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Reset a password from a mailed link
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Rotate a refresh token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      parameters:
      - description: Email and optional locale
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      responses:
        "204":
          description: accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "403":
          description: registration disabled
          schema:
            $ref: '#/definitions/api.errorBody'
      summary: Request account enrollment
      tags:
      - auth
...
  /api/v1/contexts:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Context'
            type: array
      security:
      - BearerAuth: []
      summary: List contexts
      tags:
      - contexts
    post:
      parameters:
      - description: Context
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.contextRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Context'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Create a context
      tags:
      - contexts
  /api/v1/contexts/{id}:
    delete:
      parameters:
      - description: Context id
        in: path
        name: id
        required: true
        type: integer
      - description: Delete even if it holds actions
        in: query
        name: force
        type: boolean
      responses:
        "204":
          description: deleted
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.contextInUseBody'
      security:
      - BearerAuth: []
      summary: Delete a context
      tags:
      - contexts
    get:
      parameters:
      - description: Context id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Context'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Get a context
      tags:
      - contexts
    put:
      parameters:
      - description: Context id
        in: path
        name: id
        required: true
        type: integer
      - description: Context
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.contextRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Context'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Update a context
      tags:
      - contexts
  /api/v1/export:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Export my data
      tags:
      - transfer
  /api/v1/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Get the current account
      tags:
      - account
  /api/v1/me/deletion:
    post:
      responses:
        "204":
          description: confirmation email sent
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Request deletion of my account
      tags:
      - account
  /api/v1/me/email-change:
    post:
      parameters:
      - description: Proposed new email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.emailChangeRequest'
      responses:
        "204":
          description: verification email sent
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Request a change to my email address
      tags:
      - account
  /api/v1/me/password:
    post:
      parameters:
      - description: New password plus proof of identity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.changePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.TokenPair'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Change my password
      tags:
      - account
  /api/v1/me/reauth/passkey/begin:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Begin a passkey re-authentication
      tags:
      - account
  /api/v1/notes:
    get:
      parameters:
      - description: Filter by project id
        in: query
        name: project
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Note'
            type: array
      security:
      - BearerAuth: []
      summary: List notes
      tags:
      - notes
    post:
      parameters:
      - description: Note
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.noteRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Note'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Create a note
      tags:
      - notes
  /api/v1/notes/{id}:
    delete:
      parameters:
      - description: Note id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: deleted
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Delete a note
      tags:
      - notes
    put:
      parameters:
      - description: Note id
        in: path
        name: id
        required: true
        type: integer
      - description: Note
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.noteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Note'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorBody'
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
  /api/v1/passkeys:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Credential'
            type: array
      security:
      - BearerAuth: []
      summary: List my passkeys
      tags:
      - passkeys
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"